LayoutTests/fast/canvas/canvas-pattern-transform-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS function () {
                ctx.setTransform({a: 1, m11: 11, b: 2, m12: 12, c: 3, m21: 21, d: 4, m22: 22, e: 5, m41: 41, f: 6, m42: 42});
            } threw exception TypeError: init.a and init.m11 do not match.

Testing canvas repeat transform  pattern transform
Checking point 4 4
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 124 124
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 64 32
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 64 96
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 32 64
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 96 64
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 32 32
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 96 32
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 96 32
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 96 96
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue

Testing canvas no-repeat transform  pattern transform scale(0.5)
Checking point 2 2
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 31 31
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 6 6
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 16 16
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 64 64
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 96 96
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue

Testing canvas repeat transform  pattern transform rotate(45deg)
Checking point 2 2
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 42 2
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 42 64
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 32 32
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 96 96
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue

Testing canvas repeat transform  pattern transform translate(32px, 0)
Checking point 4 4
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 124 124
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 64 32
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 2 96
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue

Testing canvas repeat transform  pattern transform translate(32px, 32px) rotate(45deg) scale(0.5)
Checking point 24 24
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 104 104
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 22 64
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 90 112
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 16 32
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 82 58
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue

Testing canvas repeat transform scale(0.5) pattern transform rotate(45deg)
Checking point 4 4
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 62 62
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 2 24
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 24 2
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 24 2
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 32 48
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 66 4
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 4 66
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
Checking point 104 104
PASS imgdata[0] is expectedRed
PASS imgdata[1] is expectedGreen
PASS imgdata[2] is expectedBlue
